Rename actions.ts to actionCreators.ts
