+++ client/react/app/actions/index.ts
@@ -1,0 +1,7 @@
+import * as actionCreators from './actionCreators';
+import * as actionConstants from './actionConstants';
+
+export default {
+  actionCreators,
+  actionConstants
+};
